requirements.txt
asgiref
pytz
sqlparse
psycopg2
redis
celery
elasticsearch
elasticsearch-dsl 
Django
django-extensions
django-filter
djangorestframework
djangorestframework-jsonapi
django-elasticsearch-dsl
django-elasticsearch-dsl-drf
python-logstash
kibana
django-prometheus
prometheus-client
